WIP refactoring event consumption
Moved EventConsumer to Consumers context
Updated Gateway to accept handlers configuration
Updated Sharding to accept handlers configuration
Removed (now) obsolete docs from Caching
Removed (now) obsolete docs from EventProducer
Removed (now) obsolete docs from EventConsumer
Added Consumers context which will add the correct consumers for a handler
Updated iex DummyHandler code
